Refactor(eos_designs): Create structured_config_utils module
